fix: switches from js to ts and adds improved error handling
